requirements.txt
fastapi
uvicorn
psycopg2-binary
pika
python-consul
requests
pyjwt

